+++ mainx.py
@@ -5,8 +5,10 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import declarative_base, sessionmaker, Session
 from sqlalchemy import Boolean, Column, Float, String, Integer
 
+from api_web import place_api
 
 app = FastAPI()
+app.include_router(place_api.router)
 
 #SqlAlchemy Setup
 SQLALCHEMY_DATABASE_URL = 'sqlite+pysqlite:///.db.sqlite3:'
@@ -70,24 +72,6 @@ def create_place(db: Session, place: Place):
     return db_place
 
 
-@app.post('/places/', response_model=Place)
-def create_places_view(place: Place, db: Session = Depends(get_db)):
-    db_place = create_place(db, place)
-    return db_place
-
-
-@app.get('/places/', response_model=List[Place])
-def get_places_view(db: Session = Depends(get_db)):
-    return get_places(db)
-
-
-@app.get('/place/{place_id}')
-def get_place_view(place_id: int, db: Session = Depends(get_db)):
-    return get_place(db, place_id)
-
-
 @app.get('/')
 async def root():
     return {'message': 'Hello World!'}
-
-
